requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cupy-cuda115==9.6.0
fastrlock==0.8
filelock==3.4.2
huggingface-hub==0.2.1
idna==3.3
joblib==1.1.0
nltk==3.6.7
numpy==1.21.4
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.6
PyYAML==6.0
regex==2021.11.10
requests==2.27.1
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
transformers==4.15.0
typing_extensions==4.0.1
urllib3==1.26.8
